+++ packages/vue/vite.config.ts
@@ -1,9 +1,17 @@
 import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
+import dts from 'vite-plugin-dts';
 import { resolve } from 'path';
 
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [
+    vue(),
+    dts({
+      outDir: 'dist/types',
+      insertTypesEntry: true,
+      rollupTypes: true,
+    }),
+  ],
   css: {
     preprocessorOptions: {
       scss: {
@@ -26,12 +34,18 @@ export default defineConfig({
       fileName: 'index',
     },
     rollupOptions: {
-      external: ['vue', '@grundtone/core', '@grundtone/utils'],
+      external: [
+        'vue',
+        '@grundtone/core',
+        '@grundtone/utils',
+        '@grundtone/icons',
+      ],
       output: {
         globals: {
           vue: 'Vue',
           '@grundtone/core': 'GrundtoneCore',
           '@grundtone/utils': 'GrundtoneUtils',
+          '@grundtone/icons': 'GrundtoneIcons',
         },
       },
     },