+++ apps/playground/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @grundtone/vue-playground
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @grundtone/icons@2.22.0
+  - @grundtone/vue@2.22.0
+
 ## 0.0.28
 
 ### Patch Changes